% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_function_template.R
\name{aaa_function_template}
\alias{aaa_function_template}
\title{Say what this function does in one line}
\usage{
aaa_function_template(data, x)
}
\arguments{
\item{data}{is the data}

\item{x}{is a variable representing something}
}
\description{
Say what this function does in greater detail if that helps.
}
\examples{
aaa_function_template(mtcars, x = mpg)
}
\keyword{efficiency}
\keyword{organization}
