% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_file_paths.R
\name{add_file_paths}
\alias{add_file_paths}
\title{adds a standardized file_paths.R project root directory}
\usage{
add_file_paths(dir_R = here::here())
}
\description{
This function adds a standardized file_paths.R project root directory
}
\examples{
add_file_paths()
Overview: Source this script to load all file path strings for your project
}
\keyword{file_paths.R}
