% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gitignore.R
\name{add_gitignore}
\alias{add_gitignore}
\title{adds a standardized gitignore file to project root directory}
\usage{
add_gitignore(dir_R = here::here())
}
\description{
This function adds a standardized gitignore file to project root directory
}
\examples{
add_gitignore()
}
\keyword{gitignore}
