% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_makefile.R
\name{add_makefile}
\alias{add_makefile}
\title{This function adds a template Makefile to your prject's root directory}
\usage{
add_makefile()
}
\description{
This function adds a template Makefile to your prject's root directory
}
\examples{
add_makefile()
}
\keyword{make}
\keyword{makefile}
