% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_readme.R
\name{add_readme}
\alias{add_readme}
\title{adds a standardized README file to project root directory}
\usage{
add_readme(learning = TRUE, dir_R = here::here())
}
\arguments{
\item{learning}{a boolean to indicate if the README should describe the use of the vignette subdirectories}
}
\description{
This function adds a standardized README to project root directory.
}
\examples{
add_readme()
}
\keyword{README}
