% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_round.R
\name{base_round}
\alias{base_round}
\title{Say what this function does in one line}
\usage{
base_round(x, base)
}
\arguments{
\item{x}{is a variable representing something}

\item{data}{is the data}
}
\description{
Say what this function does in greater detail if that helps.
}
\examples{
aaa_function_template(mtcars, x = mpg)
}
\keyword{efficiency}
\keyword{organization}
