% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Function to convert dataframe names to lowercase without white space or punctuation besides underscores.}
\usage{
clean_names(data)
}
\arguments{
\item{data}{your tibble or dataframe}
}
\description{
Function to convert dataframe names to lowercase without white space or punctuation besides underscores.
}
\examples{
mydata <- clean_names(mydata)
}
\keyword{clean}
\keyword{names}
