% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_strings.R
\name{clean_strings}
\alias{clean_strings}
\title{Function to convert astring such as the names of a dataset to lowercase without white space or punctuation besides underscores.}
\usage{
clean_strings(x)
}
\arguments{
\item{x}{are a string such as the names of a dataset}
}
\description{
Function to convert astring such as the names of a dataset to lowercase without white space or punctuation besides underscores.
}
\examples{
clean_strings("Get RID - of loosE---stuff")
}
\keyword{clean_strings}
