% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_learning_directories.R
\name{create_learning_directories}
\alias{create_learning_directories}
\title{Creates the directories for a typical learning project}
\usage{
create_learning_directories(...)
}
\description{
Creates the directories for a typical learning project
}
\examples{
create_learning_directories()
}
\keyword{learning}
\keyword{organization}
\keyword{workflow}
