% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project_directories.R
\name{create_project_directories}
\alias{create_project_directories}
\title{Creates the directories for a typical EDA project}
\usage{
create_project_directories()
}
\arguments{
\item{project_name}{a string representing the name of the Rproject}
}
\description{
Creates the directories for a typical EDA project
}
\examples{
create_project_directories()
}
\keyword{eda}
\keyword{organization}
\keyword{workflow}
