% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subproject.R
\name{create_subproject}
\alias{create_subproject}
\title{Creates the standard set of subdirectories for a typical R project}
\usage{
create_subproject(subproject, add_makefile = TRUE, ...)
}
\arguments{
\item{subproject}{string indicating the name of the new subproject directory}
}
\description{
Creates the standard set of subdirectories for a typical R project
}
\examples{
create_subproject("my_great_hypothesis")
}
\keyword{eda}
\keyword{organization}
\keyword{workflow}
