% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pval.R
\name{format_pval}
\alias{format_pval}
\title{Formats p-values to APA standards}
\usage{
format_pval(pval)
}
\arguments{
\item{pval}{is a double representing the pvalue}
}
\description{
SFormats p-values to APA standards
}
\examples{
format_pval(.230456)
}
\keyword{p-valeus}
