% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_template.R
\name{function_template}
\alias{function_template}
\title{This function generates an R file with a roxygen2-friendly template}
\usage{
function_template(
  function_name = "my_function_template",
  dir_R = here::here("R")
)
}
\arguments{
\item{function_name}{is the name of the function you are creating}

\item{dir_R}{is the directory in which you want tosave the output R file}

\item{data}{is the data}

\item{x}{is a variable representing something}
}
\description{
This function generates an R file with a roxygen2-friendly template that will enable all metadata about your function to be read into the package documentation when you build your package
}
\examples{
function_template("my_function")

Say what this function does in greater detail if that helps.
my_function_template(mtcars, x = mpg)
}
\keyword{efficiency}
\keyword{function}
\keyword{organization}
\keyword{template}
