% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb.R
\name{gb}
\alias{gb}
\title{Returns size of object in gigabytes}
\usage{
gb(x)
}
\arguments{
\item{x}{is a data object}
}
\description{
Returns size of object in gigabytes
}
\examples{
gb(mtcars)
}
\keyword{efficiency}
\keyword{organization}
