% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_dt.R
\name{glimpse_dt}
\alias{glimpse_dt}
\title{wrapper function to enable glimpse() of data.tables.}
\usage{
glimpse_dt(data_table)
}
\arguments{
\item{data_table}{a data.table}
}
\description{
wrapper function to enable glimpse() of data.tables.
}
\examples{
glimpse_dt(data_table)
}
\keyword{efficiency}
\keyword{organization}
