% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_one.R
\name{keep_one}
\alias{keep_one}
\title{function to save ove row of data for each unique value of the ID variable}
\usage{
keep_one(data, x, seed = 57)
}
\arguments{
\item{data}{is the data}

\item{x}{is a variable, usually an ID}

\item{seed}{a seed for consistent randomization}
}
\description{
function to save ove row of data for each unique value of the ID variable. Useful when you want to be sure that you have only one row per level of your ID.
}
\examples{
mtcars \%>\% keep_one(cyl)
}
\keyword{EDA}
