% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limitless_breaks.R
\name{limitless_breaks}
\alias{limitless_breaks}
\title{Sets breaks in ggplot axis without regard to min and max limits of that axis}
\usage{
limitless_breaks(.x, .step = 1)
}
\arguments{
\item{.x}{the axis (usualy left blank in ggplot usage)}

\item{.step}{the interval step for your breaks}
}
\description{
Sets breaks in ggplot axis without regard to min and max limits of that axis
}
\examples{
ggplot(mtcars, aes(x=wt,y=mpg)) + geom_point() + scale_y_continuous(breaks = limitless_breaks)
}
\keyword{breaks}
\keyword{ggplot}
