% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_odds_to_prob.R
\name{log_odds_to_prob}
\alias{log_odds_to_prob}
\title{Converts log odds to probability value}
\usage{
log_odds_to_prob(log_odds)
}
\arguments{
\item{log_odds}{is the predicted log odds for a sample}
}
\description{
Converts log odds to probability value, which is useful for interpreting results of logistic regression.
}
\examples{
log_odds_to_prob(3.56)
}
\keyword{logistic}
\keyword{regression}
