% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{return the mode from a vector}
\usage{
Mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{is a vector}

\item{na.rm}{a logical value indicating whether to consider NAs as potential modal value.}
}
\description{
return the mode from a vector
}
\examples{
Mode(mtcars$am)
}
\keyword{mode}
