% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentize.R
\name{percentize}
\alias{percentize}
\title{converts decimal doubles (e.g., 0.98) to character values with percent symbol}
\usage{
percentize(x, digits = 0)
}
\arguments{
\item{x}{is a vector of decimal values}

\item{digits}{is the number of digits to display in the output}
}
\description{
converts decimal doubles (e.g., 0.98) to character values with percent symbol
}
\examples{
percentize(c(0.98, 0.50, 0.879))
}
\keyword{percent}
