% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_n.R
\name{separate_n}
\alias{separate_n}
\title{A wrapper version of tidyr::separate() that automatically creates and names correct number of new variables}
\usage{
separate_n(data, x, sep, fill = "right", extra = "warn")
}
\arguments{
\item{data}{is the data}

\item{x}{is a variable being separated}

\item{sep}{is the separator character}

\item{fill}{default argument passed to tidyr::separate()}

\item{extra}{default argument passed to tidyr::separate()}

\item{prefix}{prefix used in names of variables created by separation}
}
\description{
A wrapper version of tidyr::separate() that automatically creates and names correct number of new variables
}
\examples{
mtcars \%>\% rownames_to_column("model") \%>\% separate_n(x = model, sep = " ")
}
\keyword{separate}
