% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_directory.R
\name{source_directory}
\alias{source_directory}
\title{Source all files in a directory.}
\usage{
source_directory(dir)
}
\arguments{
\item{dir}{a path to a directory}
}
\description{
Source all files in a directory.
}
\examples{
source_directory(here::here("R"))
}
\keyword{efficiency}
\keyword{organization}
