% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_file_paths.R
\name{source_file_paths}
\alias{source_file_paths}
\title{function to source file paths script}
\usage{
source_file_paths(file_paths = here::here("file_paths.R"))
}
\arguments{
\item{file_paths}{is an Rscript of files paths}
}
\description{
function to source file paths script where I save all my file paths.
}
\examples{
source_file_paths()
}
\keyword{file}
\keyword{paths}
