% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_rmd.R
\name{source_rmd}
\alias{source_rmd}
\title{source() Rmd files}
\usage{
source_rmd(rmd_file)
}
\arguments{
\item{rmd_file}{is a string indicating the file or file path}
}
\description{
This function would source() .Rmd files, which can be useful in makelikefiles
}
\examples{
source_rmd(here::here("munge", "100_init_munge.Rmd"))
}
\keyword{here}
\keyword{makelikefile}
