% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_f003.R
\name{temp_f003}
\alias{temp_f003}
\title{Say what this function does in one line}
\usage{
temp_f003(data, x)
}
\arguments{
\item{data}{is the data}

\item{x}{is a variable representing something}
}
\description{
Say what this function does in greater detail if that helps.
}
\examples{
temp_f003(mtcars, x = mpg)
}
\keyword{efficiency}
\keyword{organization}
