% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_foo.R
\name{temp_foo}
\alias{temp_foo}
\title{Say what this function does in one line}
\usage{
temp_foo(name, birthdate)
}
\arguments{
\item{data}{is the data}

\item{x}{is a variable representing something}
}
\description{
Say what this function does in greater detail if that helps.
}
\examples{
temp_foo(mtcars, x = mpg)
}
\keyword{efficiency}
\keyword{organization}
