% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{wrap}
\alias{wrap}
\title{inserts line breaks into a string value after number of characters in `width`}
\usage{
wrap(x, width = 20)
}
\arguments{
\item{x}{a string value}
}
\description{
inserts line breaks into a string value after number of characters in `width`. This is useful in ensuring that ggplot titles do not run off the panel.
}
\examples{
wrap("This is a really long title of a plot that I want to nicely wrap and fit onto the plot without having to manually add the backslash n, but at the moment it does not", width = 20)
}
\keyword{ggplot}
\keyword{title}
\keyword{wrap}
