% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_round.R
\name{base_round}
\alias{base_round}
\title{A function to round values to any base}
\usage{
base_round(x, base)
}
\arguments{
\item{x}{is the vector you will be rounding}

\item{base}{is the value you will be rounding to (e.g., 25 or .25)}
}
\description{
This function will round values to any base (e.g., 25 or .25)
}
\examples{
base_round( c(23.67, 2.7, 52.9, 7.7, 99.4), base = .25 )
base_round( c(23.67, 2.7, 52.9, 7.7, 99.4), base = 25 )
}
\keyword{base}
\keyword{round}
