% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project_directories.R
\name{create_project_directories}
\alias{create_project_directories}
\title{Creates the standard set of subdirectories for a typical R project}
\usage{
create_project_directories(
  style = "default",
  license = "none",
  vignette = FALSE,
  add_makefile = TRUE,
  ...
)
}
\arguments{
\item{style}{string value indicating the style of subdirectories to create (c("default", "learning", "minimal"))}

\item{vignette}{a boolean value indicating whether "analysis" directory should be named "vignette"}
}
\description{
Creates the standard set of subdirectories for a typical R project
}
\examples{
create_project_directories()
}
\keyword{eda}
\keyword{organization}
\keyword{workflow}
